Auth fixed + Customer CRUD API finished

+++ next-door-farm-backend/next-door-farm-backend/Models/Orders.cs
@@ -8,7 +8,7 @@ namespace next_door_farm_backend.Models
         [Key]
         public Guid RefID { get; set; }
         public string Status { get; set; }
-        public float Rating { get; set; }
+        public double Rating { get; set; }
         public Guid CustomerId { get; set; }
 
         public Orders()
