+++ next-door-farm-backend/next-door-farm-backend/Models/Products.cs
@@ -9,7 +9,7 @@ namespace next_door_farm_backend.Models
         public Guid RefID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public float PricePerKg { get; set; }
+        public double PricePerKg { get; set; }
         public string ImageLink { get; set; }
         public int Amount { get; set; }
         public Guid FarmerId { get; set; }
